Move the ItemLookupError into a separate file
