+++ common/stacks/src/main/scala/weco/api/stacks/http/SierraItemLookupError.scala
@@ -1,0 +1,14 @@
+package weco.api.stacks.http
+
+import weco.api.stacks.models.SierraErrorCode
+
+sealed trait SierraItemLookupError
+
+object SierraItemLookupError {
+  case class ItemHasNoStatus(t: Throwable) extends SierraItemLookupError
+
+  case object ItemNotFound extends SierraItemLookupError
+
+  case class UnknownError(errorCode: SierraErrorCode)
+    extends SierraItemLookupError
+}
